decoupled core instagram functionality from authentication and signing to support unauthenticate proxy use cases
